correct dbservice endpoint and add health checks
correct dbservice endpoint and add health checks

+++ tform/simple-microservice/infra/clean_start.sh
@@ -1,10 +1,20 @@
 #!/bin/bash
 
 echo 'Cleaning previous state'
-rm -rf terraform.tfstate* .terraform*
-rm -rf *.pem 
 
-terraform init
+read -p "Are you sure? " -n 1 -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    cp terraform.tfstate.state terraform.tfstate.old
 
-./scripts/import_domain.sh
-terraform apply
+    rm -rf terraform.tfstate* .terraform*
+    rm -rf *.pem 
+
+    terraform init
+
+    ./scripts/import_domain.sh
+    terraform apply
+fi
+
+
